feat: implement annual leave rules change & activity proof image upload with on-hold status

+++ backend/app/Models/Employee.php
@@ -21,7 +21,7 @@ class Employee extends Model
 
     protected $casts = [
         'joined_at' => 'date:Y-m-d',
-        'leave_balance' => 'integer',
+        'leave_balance' => 'float',
     ];
 
     /**
